feels like this should just be a reeal rest api

+++ utils.sh
@@ -91,11 +91,13 @@ read -p "Type a Command: " cmd
         rmusr)
             USER_ID=$(mysql -uroot -proot jream -e "SELECT id FROM user WHERE email = 'testbot01@jream-ignore.com'" | grep -o '[0-9]*');
             while true; do
-              mysql -uroot -proot jream -e "DELETE FROM \`user\` WHERE \`alias\` = \"testbotone\""
+              mysql -uroot -proot jream -e "DELETE FROM \`user\` WHERE \`alias\` = \"testbotone\" OR \"email\"='testbot01@jream-ignore.com'"
               mysql -uroot -proot jream -e "DELETE FROM newsletter_subscription WHERE \`email\` = \"testbot01@jream-ignore.com\""
-              mysql -uroot -proot jream -e "DELETE FROM user_purchase  WHERE \`user_id\` = \"${USER_ID}\""
-              mysql -uroot -proot jream -e "DELETE FROM transaction  WHERE \`user_id\` = \"${USER_ID}\""
-              echo "DELETE from user_purchase WHERE user_id = ${USER_ID}"
+              if [ -z $USER_ID ]; then
+                mysql -uroot -proot jream -e "DELETE FROM user_purchase  WHERE \`user_id\` = \"${USER_ID}\""
+                mysql -uroot -proot jream -e "DELETE FROM transaction  WHERE \`user_id\` = \"${USER_ID}\""
+                USER_ID=''
+              fi
               sleep 10
             done
             ;;
